# go-core/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateMistakeRequest:
    properties:
      category:
        $ref: '#/definitions/data.MistakeCategory'
      name:
        maxLength: 255
        minLength: 1
        type: string
      user_id:
        type: integer
    required:
    - category
    - name
    - user_id
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  dto.GetMistakesResponse:
    properties:
      mistakes:
        items:
          $ref: '#/definitions/dto.MistakeResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
  dto.GetRulesResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      rules:
        items:
          $ref: '#/definitions/dto.RuleResponse'
        type: array
    type: object
  dto.GetStrategiesResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      strategies:
        items:
          $ref: '#/definitions/dto.StrategyResponse'
        type: array
    type: object
...
  dto.MistakeResponse:
    properties:
      category:
        $ref: '#/definitions/data.MistakeCategory'
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  dto.UpdateMistakeRequest:
    properties:
      category:
        $ref: '#/definitions/data.MistakeCategory'
      id:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      user_id:
        type: integer
    required:
    - category
    - id
    - name
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive trading journal API for tracking trades, setups, and
    market analysis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 100xTrader API
  version: "1.0"
paths:
  /api/v1/mistakes:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of mistakes
      parameters:
      - description: 'Number of mistakes to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of mistakes to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mistakes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMistakesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List mistakes
      tags:
      - mistakes
    post:
      consumes:
      - application/json
      description: Create a new trading mistake
      parameters:
      - description: Mistake data
        in: body
        name: mistake
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMistakeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Mistake created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MistakeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new mistake
      tags:
      - mistakes
  /api/v1/mistakes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific mistake by ID
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mistake deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Invalid mistake ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Mistake not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a mistake
      tags:
      - mistakes
    get:
      consumes:
      - application/json
      description: Retrieve a specific mistake with all its details
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Mistake retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MistakeResponse'
              type: object
        "400":
          description: Invalid mistake ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Mistake not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a mistake by ID
      tags:
      - mistakes
    put:
      consumes:
      - application/json
      description: Update an existing mistake with new data
      parameters:
      - description: Mistake ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated mistake data
        in: body
        name: mistake
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMistakeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Mistake updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MistakeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Mistake not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a mistake
      tags:
      - mistakes
  /api/v1/mistakes/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of mistakes for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of mistakes to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of mistakes to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User mistakes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMistakesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get mistakes by user
      tags:
      - mistakes
  /api/v1/mistakes/user/{user_id}/category/{category}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of mistakes filtered by category for
        a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Mistake category
        in: path
        name: category
        required: true
        type: string
      - description: 'Number of mistakes to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of mistakes to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User mistakes by category retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetMistakesResponse'
              type: object
        "400":
          description: Invalid user ID, category, or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get mistakes by category
      tags:
      - mistakes
  /api/v1/rules:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of rules
      parameters:
      - description: 'Number of rules to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of rules to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rules retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetRulesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List rules
      tags:
      - rules
    post:
      consumes:
      - application/json
      description: Create a new trading rule
      parameters:
      - description: Rule data
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Rule created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RuleResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new rule
      tags:
      - rules
  /api/v1/rules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rule deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Invalid rule ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Rule not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a rule
      tags:
      - rules
    get:
      consumes:
      - application/json
      description: Retrieve a specific rule with all its details
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rule retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RuleResponse'
              type: object
        "400":
          description: Invalid rule ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Rule not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a rule by ID
      tags:
      - rules
    put:
      consumes:
      - application/json
      description: Update an existing rule with new data
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated rule data
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rule updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RuleResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Rule not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a rule
      tags:
      - rules
  /api/v1/rules/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of rules for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of rules to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of rules to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User rules retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetRulesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get rules by user
      tags:
      - rules
  /api/v1/rules/user/{user_id}/category/{category}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of rules filtered by category for a specific
        user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Rule category
        in: path
        name: category
        required: true
        type: string
      - description: 'Number of rules to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of rules to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User rules by category retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetRulesResponse'
              type: object
        "400":
          description: Invalid user ID, category, or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get rules by category
      tags:
      - rules
  /api/v1/strategies:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of strategies
      parameters:
      - description: 'Number of strategies to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of strategies to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Strategies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetStrategiesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List strategies
      tags:
      - strategies
    post:
      consumes:
      - application/json
      description: Create a new trading strategy
      parameters:
      - description: Strategy data
        in: body
        name: strategy
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStrategyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Strategy created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StrategyResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new strategy
      tags:
      - strategies
  /api/v1/strategies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific strategy by ID
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Strategy deleted successfully
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Invalid strategy ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Strategy not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a strategy
      tags:
      - strategies
    get:
      consumes:
      - application/json
      description: Retrieve a specific strategy with all its details
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Strategy retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StrategyResponse'
              type: object
        "400":
          description: Invalid strategy ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Strategy not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a strategy by ID
      tags:
      - strategies
    put:
      consumes:
      - application/json
      description: Update an existing strategy with new data
      parameters:
      - description: Strategy ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated strategy data
        in: body
        name: strategy
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStrategyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Strategy updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StrategyResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Strategy not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a strategy
      tags:
      - strategies
  /api/v1/strategies/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of strategies for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of strategies to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of strategies to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User strategies retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetStrategiesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get strategies by user
      tags:
      - strategies
  /api/v1/trades:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of trades
      parameters:
      - description: 'Number of trades to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of trades to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trades retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetTradesResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List trades
      tags:
      - trades
    post:
      consumes:
      - application/json
      description: Create a new trade with all required and optional fields
      parameters:
      - description: Trade data
        in: body
        name: trade
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTradeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Trade created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TradeResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new trade
      tags:
      - trades
...
  /api/v1/trades/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of trades for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 'Number of trades to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of trades to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User trades retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetTradesResponse'
              type: object
        "400":
          description: Invalid user ID or query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get trades by user
      tags:
      - trades
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users
      parameters:
      - description: 'Number of users to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of users to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetUsersResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new user
      tags:
      - users
...
schemes:
- http
swagger: "2.0"